added more information when memcached sends back "invalid" responses.

+++ util/build.sh
@@ -2,7 +2,7 @@
 
 # this file is mostly meant to be used by the author himself.
 
-#ragel -I src -G2 src/ngx_http_memc_response.rl
+ragel -I src -G2 src/ngx_http_memc_response.rl
 
 if [ $? != 0 ]; then
     echo 'Failed to generate the memcached response parser.' 1>&2
@@ -10,32 +10,38 @@ if [ $? != 0 ]; then
 fi
 
 root=`pwd`
-cd ~/work
+#cd ~/work
 version=$1
 #opts=$2
 home=~
+target=$root/work/nginx
+
+if [ ! -d ./buildroot ]; then
+    mkdir ./buildroot || exit 1
+fi
+
+cd buildroot || exit 1
+
 
 if [ ! -s "nginx-$version.tar.gz" ]; then
-    wget "http://sysoev.ru/nginx/nginx-$version.tar.gz" -O nginx-$version.tar.gz || exit 1
-    tar -xzvf nginx-$version.tar.gz || exit 1
-    if [ "$version" = "0.8.41" ]; then
-        cp $root/../no-pool-nginx/nginx-$version-no_pool.patch ./
-        patch -p0 < nginx-$version-no_pool.patch || exit 1
+    if [ -f ~/work/nginx-$version.tar.gz ]; then
+        cp ~/work/nginx-$version.tar.gz ./ || exit 1
+    else
+        wget "http://sysoev.ru/nginx/nginx-$version.tar.gz" -O nginx-$version.tar.gz || exit 1
     fi
+
+    tar -xzvf nginx-$version.tar.gz || exit 1
 fi
 
 #tar -xzvf nginx-$version.tar.gz || exit 1
 #cp $root/../no-pool-nginx/nginx-0.8.53-no_pool.patch ./
 #patch -p0 < nginx-0.8.53-no_pool.patch
 
-if [ -n "$2" ]; then
-    cd nginx-$version-$2/
-else
-    cd nginx-$version/
-fi
+cd nginx-$version/
 
 if [[ "$BUILD_CLEAN" -eq 1 || ! -f Makefile || "$root/config" -nt Makefile || "$root/util/build.sh" -nt Makefile ]]; then
-    ./configure --prefix=/opt/nginx \
+    ./configure --prefix=$target \
+        --with-cc-opt="-O2" \
           --with-http_addition_module \
             --without-mail_pop3_module \
             --without-mail_imap_module \
@@ -51,17 +57,17 @@ if [[ "$BUILD_CLEAN" -eq 1 || ! -f Makefile || "$root/config" -nt Makefile || "$
           --add-module=$root/../ndk-nginx-module \
           --add-module=$root/../eval-nginx-module \
           --add-module=$root/../echo-nginx-module \
-          --add-module=$home/work/nginx/ngx_http_upstream_keepalive-2ce9d8a1ca93 \
-          --with-debug
+          --add-module=$home/work/nginx/ngx_http_upstream_keepalive-2ce9d8a1ca93
+          #--with-debug
           #--add-module=$home/work/nginx/nginx_upstream_hash-0.3 \
   #--without-http_ssi_module  # we cannot disable ssi because echo_location_async depends on it (i dunno why?!)
 
 fi
-if [ -f /opt/nginx/sbin/nginx ]; then
-    rm -f /opt/nginx/sbin/nginx
+if [ -f $target/sbin/nginx ]; then
+    rm -f $target/sbin/nginx
 fi
-if [ -f /opt/nginx/logs/nginx.pid ]; then
-    kill `cat /opt/nginx/logs/nginx.pid`
+if [ -f $target/logs/nginx.pid ]; then
+    kill `cat $target/logs/nginx.pid`
 fi
 make -j3
 make install
